Move the SpanSnapshot into its own file
