+++ sdk/trace/span.go
@@ -593,38 +593,3 @@ type Status struct {
 	// applicable when Code is Error.
 	Description string
 }
-
-// SpanSnapshot is a snapshot of a span which contains all the information
-// collected by the span. Its main purpose is exporting completed spans.
-// Although SpanSnapshot fields can be accessed and potentially modified,
-// SpanSnapshot should be treated as immutable. Changes to the span from which
-// the SpanSnapshot was created are NOT reflected in the SpanSnapshot.
-type SpanSnapshot struct {
-	SpanContext trace.SpanContext
-	Parent      trace.SpanContext
-	SpanKind    trace.SpanKind
-	Name        string
-	StartTime   time.Time
-	// The wall clock time of EndTime will be adjusted to always be offset
-	// from StartTime by the duration of the span.
-	EndTime    time.Time
-	Attributes []attribute.KeyValue
-	Events     []Event
-	Links      []trace.Link
-	Status     Status
-
-	// DroppedAttributeCount contains dropped attributes for the span itself.
-	DroppedAttributeCount int
-	DroppedEventCount     int
-	DroppedLinkCount      int
-
-	// ChildSpanCount holds the number of child span created for this span.
-	ChildSpanCount int
-
-	// Resource contains attributes representing an entity that produced this span.
-	Resource *resource.Resource
-
-	// InstrumentationLibrary defines the instrumentation library used to
-	// provide instrumentation.
-	InstrumentationLibrary instrumentation.Library
-}
